6-size.c

+++ 0x00-hello_world/6-size.c
@@ -13,8 +13,8 @@ int main(void)
 	char c;
 	float f;
 	
-	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("Size of an int: %ld byte(s)\n", sizeof(i));
-	printf("size of a long int: %ld byte(s)\n", sizeof(li));
-	printf("size of a long long int: %ld byte(s)\n", sizeof(i));
-	printf("size of a float: %ld byte(s)\n", sizeof(f));
+	printf('Size of a char: %lu byte(s)\n', sizeof(c));
+	printf('Size of an int: %ld byte(s)\n', sizeof(i));
+	printf('size of a long int: %ld byte(s)\n', sizeof(li));
+	printf('size of a long long int: %ld byte(s)\n', sizeof(i));
+	printf('size of a float: %ld byte(s)\n', sizeof(f));
